+++ app/__init__.py
@@ -19,10 +19,20 @@ def create_app():
                         charset='utf8mb4',
                         cursorclass=pymysql.cursors.DictCursor)
 
-    from .auth import auth as auth_blueprint
-    from .main import main as main_blueprint
+    from .blueprints.auth.auth import auth as auth_blueprint
+    from .blueprints.auth.authStaff import auth as authStaff_blueprint
+    from .blueprints.auth.authCust import auth as authCust_blueprint
+
+    from .blueprints.main.main import main as main_blueprint
+    from .blueprints.main.mainCust import main as mainCust_blueprint
+    from .blueprints.main.mainStaff import main as mainStaff_blueprint
+
     app.register_blueprint(auth_blueprint)
+    app.register_blueprint(authStaff_blueprint)
+    app.register_blueprint(authCust_blueprint)
     app.register_blueprint(main_blueprint)
+    app.register_blueprint(mainCust_blueprint)
+    app.register_blueprint(mainStaff_blueprint)
 
     return app
